fixup! feat: implemented useUnlocks hook
