+++ tests/constants/votes.js
@@ -30,53 +30,4 @@ const generateVotes = (index) => ({
   height: 16418742 + index,
 });
 
-const generateVotesWrapper = (votes, extraParams = {}) => ({
-  votes,
-  account: {
-    address: 'lsk24cd35u4jdq8szo3pnsqe5dsxwrnazyqqqg5eu',
-    publicKey: 'aq02qkbb35u4jdq8szo3pnsqe5dsxwrnazyqqqg5eu',
-    name: 'genesis_56',
-    ...extraParams,
-  },
-});
-
-const generateUnlocksWrapper = (unlocking, extraParams = {}) => ({
-  unlocking,
-  account: {
-    address: 'lsk24cd35u4jdq8szo3pnsqe5dsxwrnazyqqqg5eu',
-    publicKey: 'aq02qkbb35u4jdq8szo3pnsqe5dsxwrnazyqqqg5eu',
-    name: 'genesis_56',
-    ...extraParams,
-  },
-});
-
-const generateVote = (index) => ({
-  delegateAddress: `lsk24cd35u4jdq8sz${index}ptrn47dsxwrnazyhhkg5eu`,
-  amount: `1000${index}`,
-  name: 'liskhq',
-});
-
-const generateUnlock = (index) => ({
-  delegateAddress: `lsk24cd35u4jdq8szo3pnsqe5dsxwrnazyqqqg${index}eu`,
-  amount: `100${index}`,
-  unvoteHeight: {
-    start: 30 + index,
-    end: 300 + index,
-  },
-}
-);
-
 export const votesList = Array(10).fill(1).map((_, idx) => generateVotes(idx));
-
-export const sentVotesList = generateVotesWrapper(
-  Array(30).fill(1).map((_, idx) => generateVote(idx)),
-  { votesUsed: 10 },
-);
-
-export const receivedVotesList = generateVotesWrapper(
-  Array(30).fill(1).map((_, idx) => generateVote(idx)),
-);
-
-export const unlockList = generateUnlocksWrapper(
-  Array(30).fill(1).map((_, idx) => generateUnlock(idx)),
-);
